Added detail page routing to AdminX demo app
